+++ web-component/lib/render.ts
@@ -1,14 +1,14 @@
 export function renderHTML():string{
   return /* html */ `
-    <div class="jb-input-web-component">
+    <div class="jb-input-web-component" part="root">
     <label for="input" part="label"></label>
-    <div class="input-box" part="input-box">
-        <div class="jb-input-start-section-wrapper">
-            <slot name="start-section"></slot>
+    <div class="control" part="control">
+        <div class="jb-input-inline-start-wrapper">
+            <slot name="inline-start"></slot>
         </div>
         <input id="input" part="input" aria-describedby="message"/>
-        <div class="jb-input-end-section-wrapper">
-            <slot name="end-section"></slot>
+        <div class="jb-input-inline-end-wrapper">
+            <slot name="inline-end"></slot>
         </div>
     </div>
     <div id="message" class="message-box" part="message" aria-live="polite" aria-atomic="true"></div>